+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2026-08-12, Version 0.7.0 (Stable), @nzorba
+
+### Added
+
+- Feature to disable message text box and send button on quick replies based on two conditions:
+1. disableTextMessageSendingOnQuickReply flag from service which is set depending on whether service uses this feature
+2. contentType of message is set to 'QuickReply'
+
 ## 2026-05-08, Version 0.6.0 (Stable), @gregwolversonHO
 
 ### Changed